feat(chproxy): refactor OTEL telemetry
- Upgrade Go from 1.23 to 1.24
- Separate files (batch.go, buffer.go) for easier understanding
- Implement shared HTTP client with connection pooling
- Enhance telemetry with request counter and atomic buffer operations
- Add compression for log exports with minimum severity filtering
- Improve graceful shutdown handling with in-flight request tracking
- Increase default flush interval from 3s to 5s
- Bump service version from 1.1.0 to 1.2.0
